% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bic_d.R
\name{bic_d}
\alias{bic_d}
\title{Modified-BIC type criterion}
\usage{
bic_d(lambdas, n, p)
}
\arguments{
\item{lambdas}{The eigenvalues of the matrix that spans the subspace}

\item{n}{The number of observations.}

\item{p}{The}
}
\description{
\code{bic_d} estimates the dimension of a subspace.
}
\details{
The function estimates the dimension of a subspace using
the modified-BIC type criterion of Zhu et al. (2010).
The estimation is based on choosing \eqn{k} that maximizes
\eqn{G_n(k)=n \frac{\sum_{i=1}^{k} \hat{\lambda}_{i}^2}{\sum_{i=1}^{p}\hat{\lambda}_{i}^2}}
}
