% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantdr.R
\docType{package}
\name{quantdr}
\alias{quantdr}
\title{quantdr: A package for performing dimension reduction on conditional quantiles.}
\description{
The quantdr package provides two important functions:
llqr and cqs.
}
\section{quantdr functions}{

}

