% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqs.R
\name{cqs}
\alias{cqs}
\title{Central quantile subspace}
\usage{
cqs(x, y, tau = 0.5, dtau)
}
\arguments{
\item{x}{A design matrix.  The rows represent observations and the columns
represent predictor variables.}

\item{y}{A vector of the response variable.}

\item{tau}{A quantile level, a number strictly between 0 and 1.}

\item{dtau}{The dimension of the central quantile subspace.  If not
specified, \code{dtau} is estimated using the modified-BIC type criterion
of Zhu et al. (2010)}
}
\value{
\code{cqs} computes the directions of the central quantile subspace,
  and returns:
  \itemize{
  \item{qvectors: }{The estimated directions of the
  \eqn{\tau}th central quantile subspace.}

  \item{qvalues: }{The eigenvalues resulting from the eigenvalue decomposion
  of the matrix with column vectors the estimated vectors. If \code{d_tau} is
  one, the \code{qvalues} output is not produced.}

  \item{dtau: }{The dimension of the central quantile subspace.  If not
  specified by the user, \code{dtau} is estimated using the modified-BIC
  type criterion of Zhu et al. (2010).}
  }
}
\description{
\code{cqs} estimates the directions of the central quantile subspace.
}
\details{
The function computes the directions that span the \eqn{\tau}th central
quantile subspace, i.e., the directions that define linear combinations of
the predictor \code{x} that contain all the information available on the
conditional quantile function.

The function starts by estimating the initial vector, which is defined as the
least-squares estimator from regressing the conditional quantile on \code{x}.
Then, if the dimension of the central quantile subspace is one, the algorithm
stops and reports that vector as the basis of the central quantile subspace.
Otherwise, the algorithm continues by creating more vectors and applying an
eigenvalue decomposition to extract linearly independent vectors.  If the
dimension of the central quantile subspace is unknown, it is estimated using
the modified-BIC type criterion of Zhu et al. (2010).
}
\examples{
# Example 1
# estimate the directions of a single-index model
set.seed(1234)
n <- 100; p <- 10
x <- matrix(rnorm(n * p), n, p); error <- rnorm(n)
y <- 3 * x[, 1] + x[, 2] + error
tau <- 0.5
out <- cqs(x, y, tau, dtau = 1)
out
# without specifying d and dtau
out <- cqs(x, y, tau)
out
out$qvectors[, 1:out$dtau]

# Example 2
# estimate the directions of a multi-index model
set.seed(1234)
n <- 100; p <- 10
x <- matrix(rnorm(n * p), n, p); error <- rnorm(n)
y <- exp(x[, 1]) + x[, 2] + error
tau <- 0.5
cqs(x, y, tau)$qvectors[, 1:2]

}
\references{
Zhu, L.-P., Zhu, L.-X., Feng, Z.-H. (2010) Dimension reduction in
  regression through cumulative slicing estimation. \emph{Journal of the
  American Statistical Association}, 105, 1455-1466.
}
