% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llqr.R
\name{llqr}
\alias{llqr}
\alias{llqrcv}
\title{Local linear quantile estimation}
\usage{
llqr(x, y, tau = 0.5, h, method = "rule", x0)

llqrcv(x, y, tau = 0.5)
}
\arguments{
\item{x}{a design matrix.The rows represent observations and the columns
represent predictor variables}

\item{y}{a vector of the respons variable}

\item{tau}{the quantile level, a number strictly between 0 and 1.}

\item{h}{A univariate bandwidth.}

\item{method}{A method to select the bandwidth, if it is missing.  Use
\code{rule} for the rule-of-thumb bandwidth of Yu and Jones (1994) or
\code{CV} for the method of cross-validation.}

\item{x0}{A single observation for which to perform the estimation.  It needs
to be a singular value, for a univariate predictor, or a vector, for a
multivariate predictor.  If \code{x0} is missing, the estimation will be
performed on the design matrix \code{x}.}
}
\value{
\code{llqr} computes the local linear \eqn{\tau}-th conditional
  quantile estimator of \code{y} given \code{x}, and returns:

  \code{ll_est}   The estimated function value at the design points \code{x}
  or, if specified, at the point \code{x0}.

  \code{h}  The bandwidth.

\code{llqrcv} returns the optimal bandwidth selected using
  Cross-Validation criterion for the estimation of the \code{tau}-th
  conditional quantile of \code{y} given \code{x}.
}
\description{
\code{llqr} estimates the \eqn{\tau}-th conditional quantile of \code{y}
given \code{x} based on a local linear fitting.  The estimation is performed
at each of the design points or, if specified, at a single observation point.
}
\details{
The function computes the local linear quantile regression fit for a specified
quantile \eqn{\tau} at the design points of the matrix \code{x} or at a
pre-specified point \code{x0}.  The estimation is based on a standard normal
kernel and a univariate bandwidth, which, if not specified by the user, it is
defined using either the rule-of-thumb given by Yu and Jones (1994) or the
cross-validation criterion.

The estimation applies to univariate and multivariate predictor variables.
For the later, the function uses the multivariate standard normal kernel.
Note that, if the estimation is performed at a pre-specified point \code{x0},
then \code{x0} should be a scaler (for univariate predictor) or a vector (for
multivariate predictor).
}
\section{Functions}{
\itemize{
\item \code{llqrcv}: Cross-Validation for bandwidth selection for local linear
conditional quantile estimation
}}

\examples{
n <- 100
x <- rnorm(100); error <- rnorm(100); y <- x^2 + error
tau <- c(0.25, 0.5, 0.75)
par(mfrow=c(2,2))
for (i in 1:length(tau)) {
  plot(x, y, main=print(tau[i]))
  points(x, llqr(x, y, tau=tau[i])$ll_est, pch=16, col='red')
}

\dontrun{
n <- 100
x <- cbind(rnorm(n), rnorm(n)); error <- rnorm(100); y <- x[, 1]^2 + error
x0=1
tau=0.5
llqr(x, y, tau=tau, x0=x0)$ll_est
}
n <- 100
x <- rnorm(100); error <- rnorm(100); y <- x^2 + error
tau <- 0.5
llqrcv(x, y, tau = tau)
}
\references{
Yu, K. and Jones, M.C. (1998), Local linear quantile regression.
Journal of the American Statistical Association, 93, 228-237.
}
