% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llqrcv.R
\name{llqrcv}
\alias{llqrcv}
\title{Cross-Validation for bandwidth selection for local linear
conditional quantile estimation}
\usage{
llqrcv(x, y, tau = 0.5)
}
\arguments{
\item{x}{a design matrix.The rows represent observations and the columns
represent predictor variables}

\item{y}{a vector of the respons variable}

\item{tau}{the quantile level, a number strictly between 0 and 1.}
}
\value{
\code{llqrcv} returns the optimal bandwidth selected using
  Cross-Validation criterion for the estimation of the \code{tau}-th
  conditional quantile of \code{y} given \code{x}.
}
\description{
Cross-Validation for bandwidth selection for local linear
conditional quantile estimation
}
\examples{
n <- 100
x <- rnorm(100); error <- rnorm(100); y <- x^2 + error
tau <- 0.5
llqrcv(x, y, tau = tau)
}
